pearl-agent-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.4.1
openai==1.6.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
faiss-cpu==1.7.4
sentence-transformers==3.0.1
huggingface-hub==0.20.3
numpy==1.24.3
gotrue==2.12.4
